Abstracted out for structure functions
