+++ src/utils/convertPipes/index.js
@@ -1,16 +1,7 @@
 const cheerio = require("cheerio");
 const { flatMap, includes, compact } = require("lodash");
 const { unescapePiping, removeDash } = require("../HTMLUtils");
-const { unitConversion } = require("../../constants/units");
-
-const {
-  FORMAT_CURRENCY,
-  FORMAT_DATE,
-  FORMAT_NUMBER,
-  NUMBER_TRANSFORMATION,
-  DATE_TRANSFORMATION,
-  FORMAT_UNIT,
-} = require("../../constants/piping");
+const { buildStructure } = require("./transform");
 
 const getMetadata = (ctx, metadataId) =>
   ctx.questionnaireJson.metadata.find(({ id }) => id === metadataId);
@@ -20,9 +11,9 @@ const isPipeableType = (answer) => {
   return !includes(notPipeableAnswerTypes, answer.type);
 };
 
-const getAllAnswers = questionnaire =>
+const getAllAnswers = (questionnaire) =>
   flatMap(questionnaire.sections, (section) =>
-    flatMap(section.folders, folder =>
+    flatMap(section.folders, (folder) =>
       compact(flatMap(folder.pages, (page) => page.answers))
     )
   );
@@ -32,72 +23,22 @@ const getAnswer = (ctx, answerId) =>
     .filter((answer) => isPipeableType(answer))
     .find((answer) => answer.id === answerId);
 
-const DATE_FORMAT_MAP = {
-  "dd/mm/yyyy": "d MMMM yyyy",
-  "mm/yyyy": "MMMM yyyy",
-  yyyy: "yyyy",
-};
-
-const TRANSFORM_MAP = {
-  Currency: { format: FORMAT_CURRENCY, transformKey: NUMBER_TRANSFORMATION },
-  Number: { format: FORMAT_NUMBER, transformKey: NUMBER_TRANSFORMATION },
-  Date: { format: FORMAT_DATE, transformKey: DATE_TRANSFORMATION },
-  DateRange: { format: FORMAT_DATE, transformKey: DATE_TRANSFORMATION },
-  Unit: { format: FORMAT_UNIT, transformKey: NUMBER_TRANSFORMATION },
-};
-
-const transform = (AnswerType, value) => {
-  const structureOptions = (source, identifier, dateFormat, unitType) => {
-    const transformKey = [TRANSFORM_MAP[AnswerType].transformKey];
-
-    const options = {
-      [transformKey]: {
-        source,
-        identifier,
-      },
-    };
-
-    if (AnswerType === "Date") {
-      const format =
-        source === "metadata"
-          ? DATE_FORMAT_MAP["dd/mm/yyyy"]
-          : DATE_FORMAT_MAP[dateFormat];
-
-      options.date_format = format;
-    }
-
-    if (AnswerType === "Unit") {
-      options.unit = unitConversion[unitType];
-    }
-
-    return options;
-  };
-
-  const structure = {
-    value,
-    format: TRANSFORM_MAP[AnswerType].format,
-    options: structureOptions,
-  };
-
-  return structure;
-};
-
 const FILTER_MAP = {
-  Currency: (format, value) => transform(format, value),
-  Date: (format, value) => transform(format, value),
-  DateRange: (format, value) => transform(format, value),
-  Number: (format, value) => transform(format, value),
-  Unit: (format, value) => transform(format, value),
+  Currency: (format, value) => buildStructure(format, value),
+  Date: (format, value) => buildStructure(format, value),
+  DateRange: (format, value) => buildStructure(format, value),
+  Number: (format, value) => buildStructure(format, value),
+  Unit: (format, value) => buildStructure(format, value),
 };
 
 const PIPE_TYPES = {
   answers: {
     retrieve: ({ id, type }, ctx) => {
-      let tempId = id.toString()
+      let tempId = id.toString();
       if (type === "DateRange") {
-        tempId = id.endsWith("from") ? id.slice(0, -4) : id.slice(0, -2)
+        tempId = id.endsWith("from") ? id.slice(0, -4) : id.slice(0, -2);
       }
-      return getAnswer(ctx, tempId)
+      return getAnswer(ctx, tempId);
     },
     render: ({ id }) => `answer${id}`,
     getType: ({ type }) => type,
@@ -134,7 +75,11 @@ const getPipedData = (store) => (element, ctx) => {
     return "";
   }
 
-  const output = elementData.type === "DateRange" ? pipeConfig.render(elementData) : pipeConfig.render(entity);
+  const output =
+    elementData.type === "DateRange"
+      ? pipeConfig.render(elementData)
+      : pipeConfig.render(entity);
+
   const pipedType = pipeConfig.getType(entity);
 
   const transformed = FILTER_MAP[pipedType];
@@ -147,6 +92,7 @@ const getPipedData = (store) => (element, ctx) => {
 
   if (transformed) {
     let dateFormat, unitType;
+
     if (entity.properties) {
       dateFormat = entity.properties.format;
       unitType = entity.properties.unit;
@@ -158,12 +104,7 @@ const getPipedData = (store) => (element, ctx) => {
       transforms: [
         {
           transform: format,
-          arguments: options(
-            piped,
-            entity.key || output,
-            dateFormat,
-            unitType
-          ),
+          arguments: options(piped, entity.key || output, dateFormat, unitType),
         },
       ],
     };
